#47: data-holder codegen
