+++ data-holder/src/main/java/au/org/consumerdatastandards/holder/model/CommonDiscoveryStatus.java
@@ -7,20 +7,7 @@ import java.util.Objects;
 @ApiModel
 public class CommonDiscoveryStatus extends BaseResponse {
 
-    /**
-     * Get data
-     */
     private CommonDiscoveryStatusData data;
-
-    /**
-     * Get links
-     */
-    private Links links;
-
-    /**
-     * Get meta
-     */
-    private Meta meta;
 
     public CommonDiscoveryStatus data(CommonDiscoveryStatusData data) {
         this.data = data;
@@ -34,32 +21,6 @@ public class CommonDiscoveryStatus extends BaseResponse {
 
     public void setData(CommonDiscoveryStatusData data) {
         this.data = data;
-    }
-    public CommonDiscoveryStatus links(Links links) {
-        this.links = links;
-        return this;
-    }
-
-    @ApiModelProperty(required = true)
-    public Links getLinks() {
-        return links;
-    }
-
-    public void setLinks(Links links) {
-        this.links = links;
-    }
-    public CommonDiscoveryStatus meta(Meta meta) {
-        this.meta = meta;
-        return this;
-    }
-
-    @ApiModelProperty
-    public Meta getMeta() {
-        return meta;
-    }
-
-    public void setMeta(Meta meta) {
-        this.meta = meta;
     }
 
     @Override
